Fix bug when initializing AccommodationRoom from cached data.
Since `class` is a reserverd word in python, we use the attribute `room_class`
instead for the room class in AccommodationRoom model. Before, we would look
for the `class` key in the data returned by the API to populate `room_class`.
However, when locally caching the data, the room class would be serialized
under the `room_class` key, and when trying to initialize an AccommodationRoom
from that cached data, we would get a KeyError exception because `class` is not
present.