Add PromotionProduct in order to properly handle Promotion products
Promotion.products contains a list of objects that represent resources in the
API.

However, there are certain fields (`type`, `sub_type`) which are referenced
within those objects but they do not actually live within the resource.

Because gapipy defines allowed fields, those are lost upon fetch, and this
causes errors for clients which may cache, and then expect that field to
continue to exist.

The solution until the API is changed is to wrap this list of products in a
model that derives its fields from the `type` of product, and then explicitly
adds `type` and `sub_type` to its list of fields.