Properly instantiate resource lists that aren't subresources
Resources that are referred to from other resources are not inlined in their
entirety. Often, only their `id` and `href` will be inlined, and an extra API
call has to be made in order to retrieve all the data of the subresource.

In gapipy, we abstract over these details, by making all references to
resources either a Resource object, a Query object, or a plain Python list of
Resource objects. Internally, instantiated resources that aren't complete are
considered as "stubs". If a missing field is accessed and the resource is a
stub, then the full resource is fetched in the background, and the data is
returned transparently.

List of non-child resources were not properly instantiated as stubs, which
meant that accessing missing fields would raise an AttributeError, instead of
fetching it from the API.