Move cache key building from cache module to Query
This removes the need for the cache backends to know anything about the query
data. Prior to this, we were passing data required to generate a cache key to
the backend – this introduces unnecessary state. Instead the `Query` object is
now responsible for generating cache keys (see `query_key`) since it is aware
of all needed data and has always interacted with the cache backends.

This also enables us to get rid of a use of the `current_client` global
variable.

These changes were extracted from https://github.com/gadventures/gapipy/pull/20