Add dossier ids to itinerary media and highlights
A new feature has been added to gadventures linking dossiers and content
pages; however, we do not want the content pages associated with gapi.
There is a foreign key that exists between contentpages and dossiers,
but since the content pages do not exist in gapi, we need to re-establish
this relationship. We can do that by exposing the dossier_ids on itinerary-highlights.
The information already exists in gapi-layer, it simply needs to be exposed through gapipy.

As we begin to disassociate dossiers, and other structured data, from the api we'll
need to adopt this pattern of loose relationships more frequently.