Force the output of __repr__ to be of the proper string type
In Python 2, the output of `__repr__` must be a byte string, while in Python 3,
it must be unicode. Since the string representation of many G API resources
contain words in non-English languages, they can contain characters outside of
the ASCII range, which raises a UnicodeEncodeError when it is implictly coerced
to the default encoding by Python.

The way we ensure the return of a proper string type in Python 2 is by encoding
the output to UTF-8. This might not be adequate in all use cases of gapipy, but
this is what Django does in its codebase, and this seems like a good default.