Accept additional HTTP headers in `create` calls
Similar to what we do for `get` calls:
    https://github.com/gadventures/gapipy/commit/77ae9e055b26e24dd5438b5b1c1e7856c61a60ce

Basically, I'm after this functionality because I want to do some
testing of the auth ratelimiting of Trip Editor. The idea is that
gapi-layer passes the X-Forwarded-For header along to TE and TE can use
it to derive the actual client IP (to use in ratelimiting counts). The
TE auth endpoint gets hit when we attempt to *create* an oauth token.

I think that some system somewhere along the path is messing with that
header, though. Having the ability to inject my own X-Forwarded-For
header may help figure out what's happening.