Add countries.{continent,place} references
Aside: initially I had been importing the actual `Continent` and `Place`
classes to pass in as the second element of the tuples added to
`_resource_fields` ... that resulted in some circular import issues, but
poking around in the rest of the codebase I see string references as
well. From what I can tell, that seems to work here as well.