Raise an exception when `fetch`ing a stub fails (#119)
* Raise an exception when `fetch`ing a stub fails

tl;dr: This introduces better error messages for invalid stubs.

This is coming out of some discussions about this issue that has come up a number of times
sporadically in the past.

Today, gapipy swallows some HTTP errors silently (see
`gapipy.query.HTTPERRORS_MAPPED_TO_NONE`). This behaviour kind of makes
sense for the case where you are explicitly `get`ing that resource,
because you can check for that `None` condition:

    my_departure = g.departures.get(SOME_ID)
    if my_departure is None:
        pass  # or do some things

On the other hand, this behaviour is kind of confusing when you're
working with a stub that is embedded in another resource. For example if
we have a tour_dossiers instance and we want the name of the corresponding
booking company, we can do something like the following:

    bc_name = my_tour_dossier.booking_companies[0].name

If the reference is correct, this all works fine. If the reference is to
a booking_companies that does not exist, or has been deleted or to which
the client doesn't have access, though, today gapipy will:

- try and fail to fetch the `booking_companies` by its ID
- try to access the `name` attribute anyway
- throw an AttributeError, claiming that "BookingCompany has no field
name available"

This is of course, a bunch of nonsense: booking companies _do_ have
fields, we've simply failed to hydrate a broken reference.

Even if we were to do something like explicitly `fetch` that reference
before traversing, there's no way for to see the `None` that gets
returned, the stub is still a stub.

---

This PR checks for the returned None when `fetch`ing a stub's data, and
raises a more descriptive exception.

NB: It could make sense to have some custom exception type for
this instead of using ValueError. Today gapipy doesn't seem to have any
custom exceptions types, though, but it can be tabled for a future release.