Add ability to retry to Client (#124)
* Add ability to retry to Client

Add an optional argument to the gapi client that will retry failed queries.  This is useful as a workaround for the frequent timeouts TinCan is facing that result in an empty query being returned. The hope is that on the next try it will complete the query within the allotted time.

Under this hood, this creates a urllib3 `Retry` object in the requests' `HTTPAdapter`.  If someone wanted more granular control, it's totally possible to pass other params like backoff_factor to that object.

* Add env var for max_retries: `GAPI_CLIENT_MAX_RETRIES`