initial poc commit for #136 - dealing with a noop partial update
* we add the `GAPI_CLIENT_RAISE_ON_NOOP_UPDATE` environment config value
that defaults to False and store it on the Client.raise_on_noop_update
attribute.

* When attempting to make a partial update (PATCH), we compute the diff
for the data to be sent in the request against the initial `_raw_data`
value received from the G API when the Resource was initialised. The
current behaviour is that we send a PATCH request with an empty `{}`
JSON payload, and the G API will respond with a 400 error if any fields
on the resource are required. The new behaviour when there is no
computed change (i.e. data is computed to be an empty dict ({})), will
silently return a NOOP and return the unmodified Resource value to the
caller. If the Client is configured to raise an exception in the case of
a NOOP partial update, then we raise one.

It is not anticipated that this change will break existing flows for the
users of this Library; however that may still need to be ironed out.

TODO:
* add tests
* maybe: define an explicit NoopUpdateException, instead of using
ValueError