wip -- adjust tox.ini to invoke nose directly...
... and remove test-stuff from setup.py because "setup.py test" is
deprecated and we should move away from suggesting/enabling running
tests that way [0]

we'll also pass through an env var for additional nose args so that we
can dodge the "integration" tests by doing the following:

    NOSE_ARGS="-A integration!=1" tox

PLUS! at least temporarily while I'm experimenting, slim down that list
of Python versions (Keep a modern Python 3 + 2.7 for now. Before this is
ready to merge, though, we should find out which other 3.x we should be
testing. The only ones that are officially supported today are 3.7+ but
maybe we have some users that aren't up to date.)

[0] https://github.com/pypa/setuptools/issues/1684