Detect an empty payload during a partial update... (#137)
Closes #136 - dealing with a noop partial update

* we add the `GAPI_CLIENT_RAISE_ON_EMPTY_UPDATE` environment config
value that defaults to `False` and store it on the Client.raise_on_empty_update
attribute.

* When attempting to make a partial update (PATCH), we compute the diff
for the data to be sent in the request against the current `_raw_data`
value received from the G API when the Resource was initialised. The
current behaviour is that we send a PATCH request with an empty `{}`
JSON payload, and the G API will respond with a 400 error if any fields
on the resource are required. The new behaviour when there is no
computed change (i.e. data is computed to be an empty dict (`{}`)), will
silently perform a NOOP and return the unmodified Resource value to the
caller. If the Client is configured to raise an exception in the case of a
Empty partial update, it will raise the EmptyPartialUpdateError.

It is not anticipated for this change to break existing flows for the users
of this Library; however that may still need to be ironed out.

* introduce generic GapipyException and explicit EmptyPartialUpdateError
* have Resource._update raise an error when data is empty for partial
update
* Catch the exception in Resource.save and re-raise the EmptyPartialUpdateError
if the `_client.raise_on_empty_update` is `True`
* if no error is to be raised, pre-empt the return on partial-update to avoid
calling `_fill_fields` when gapipy computes an empty update payload

Co-authored-by: Jon Prindiville <jprindiville@gadventures.com>