[FIX: 110799b] the removal deepcopy(data) in _fill_fields
This was a breaking change introduced in 2.39.0 (PR #145)

The call to _fill_fields from Resource.save() updates the instance
data in place, and requires another deepcopy(data) call, which is
saved to _raw_data on the object. This is needed as a create/update
via .save() can result in additional data being returned, which needs
to be persisted onto the object to allow later .save(partial=True)
calls to correctly compute the difference in data to be sent via a
PATCH request.

This will be submitted as HOTFIX to version 2.40.0, and subsequently
2.39.0 will be yanked from PyPI.