Add optional timeout parameter to requests (#146)
Adds a new optional timeout parameter to Query.get, allowing the caller to force a timeout per call. gapipy will continue to use network configured timeout settings as it did before if no timeout is provided.

Also adds a custom timeout exception to the client: gapipy.exceptions.TimeoutError which is raised when the timeout parameter is provided.